fix(shutdown): Prevent race condition when GlobalObject destruction routine unlocks global mutex (#8652)
* fix(shutdown): Prevent race condition when ~GlobalObjectHolder() routine unlocks global mutex

Unlocking global mutex in GlobalObject destruction routine made it possible for a new attachment to slip in, so it will be creating new GlobalObject and using it, while destroying routine still in action. This can lead to an undefined state of the global objects, such as shared memory, where one thread is actively using it while another thread is destroying it.

* fix(shutdown): Reimplement synchronization between shutdown and init routine for GlobalObjectHolder

* refactor(shutdown): Remove unused variable

* refactor(shutdown): Code adjustments after Vlad review

---------

Co-authored-by: Artyom Ivanov <artyom.ivanov@red-soft.ru>